Use common voc.txt for English stemming algorithms
We were using the same 29417 entry voc.txt for "english" and "lovins",
but a different 30428 entry one for "porter".  There were a significant
number of words only in one list or the other so I've merged them to
form a single common 42603 entry voc.txt which we now use for all three
English stemming algorithms.