romanian: Expand size of test vocabulary
Back when we had romanian1 and romanian2 algorithms, both used the same
vocabulary list, but this was replaced with a different smaller one when
the algorithms were merged.

This smaller list has 26752 words (or 22570 excluding duplicates due to
having both cedilla and comma forms) which seems on the small side, so
merge in the words from the older list.  The merged list has 87644
words, or 74616 excluding duplicates.