stemmer-compare: Generate numbered lists
The merges and splits are now HTML numbered lists.  One way I review
the output is to load it into Firefox and use its machine translation
to translate to English.  With the merges and splits all numbered it
is much easier to load the output into a second browser window without
translation and quickly find the untranslated row (e.g. to look up words
in a dictionary, or to ask a native speaker).