wikipedia-dump-to-freq: Reduce junk words found
Skip special pages, which are often wholly or partly in English.

Drop links to special pages rather than expanding to the link name,
as the prefix always seems to be in English and the rest of the link
name may be too.

Add code to skip some real-world cases of broken table formatting
so we don't extract words from the table attributes in such cases.